초기세팅2
