+++ movieFront/src/main.tsx
@@ -1,22 +1,22 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import { BrowserRouter } from 'react-router-dom';
-import App from './App';
-import { ThemeProvider } from '@mui/material';
-import { theme } from './style/Theme';
+// import React from 'react';
+// import { createRoot } from 'react-dom/client';
+// import { BrowserRouter } from 'react-router-dom';
+// import App from './App';
+// import { ThemeProvider } from '@mui/material';
+// import { theme } from './style/Theme';
 
-const container = document.getElementById('root') as HTMLElement;
-const root = createRoot(container);
+// const container = document.getElementById('root') as HTMLElement;
+// const root = createRoot(container);
 
-function Root() {
-  return (
-    <BrowserRouter future={{ v7_relativeSplatPath: true, v7_startTransition: true }}>
-      <React.StrictMode>
-        <ThemeProvider theme={theme}>
-          <App />
-        </ThemeProvider>
-      </React.StrictMode>
-    </BrowserRouter>
-  );
-}
-root.render(<Root />);
+// function Root() {
+//   return (
+//     <BrowserRouter future={{ v7_relativeSplatPath: true, v7_startTransition: true }}>
+//       <React.StrictMode>
+//         <ThemeProvider theme={theme}>
+//           <App />
+//         </ThemeProvider>
+//       </React.StrictMode>
+//     </BrowserRouter>
+//   );
+// }
+// root.render(<Root />);
